requirements.txt
requests
torch
torchvision
loguru
transformers
timm
tqdm
Pillow